tc/assignments/requirements.txt
ipython
mermaid-py